pull out numpy tests to separate file
