+++ distributed/serialize/tests/test_serialize.py
@@ -9,8 +9,6 @@ from distributed.serialize.core import (register_serialization, serialize,
         deserialize, Serialize, Serialized)
 from distributed.protocol import decompress
 
-import distributed.serialize.numpy
-
 
 class MyObj(object):
     def __init__(self, data):
@@ -39,15 +37,6 @@ def test_dumps_serialize():
         result = deserialize(header, frames)
         assert result == x
 
-    x = np.ones((5, 5))
-    header, frames = serialize(x)
-    assert header['type']
-    assert len(frames) == 1
-
-    frames = decompress(header, frames)
-    result = deserialize(header, frames)
-    assert (result == x).all()
-
     x = MyObj(123)
     header, frames = serialize(x)
     assert header['type']
@@ -55,19 +44,6 @@ def test_dumps_serialize():
 
     result = deserialize(header, frames)
     assert result.data == x.data
-
-
-@pytest.mark.parametrize('x',
-        [np.ones(5),
-         np.asfortranarray(np.random.random((5, 5))),
-         np.random.random(5).astype('f4'),
-         np.empty(shape=(5, 3), dtype=[('total', '<f8'), ('n', '<f8')])])
-def test_dumps_serialize_numpy(x):
-    header, frames = serialize(x)
-    frames = decompress(header, frames)
-    y = deserialize(header, frames)
-
-    np.testing.assert_equal(x, y)
 
 
 def test_serialize_bytes():
